requirements.txt
pydantic
python-dotenv
matplotlib
openai